lab_8-9/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0  # If using Django REST Framework
psycopg2>=2.9,<3.0  # If using PostgreSQL
mysqlclient>=2.1,<3.0  # If using MySQL
pillow>=9.0,<10.0  # For image processing
django-environ>=0.10,<1.0  # For environment variable management